Rework JMXConnection Handles in JMXProxy When Node is not Managed by Local Instance
