+++ connection.impl/src/main/java/com/ericsson/bss/cassandra/ecchronos/connection/impl/providers/DistributedJmxConnectionProviderImpl.java
@@ -120,6 +120,12 @@ public class DistributedJmxConnectionProviderImpl implements DistributedJmxConne
     @Override
     public JMXConnector getJmxConnector(final UUID nodeID)
     {
+        Node node = myNativeConnectionProvider.getNodes().get(nodeID);
+        if (node == null)
+        {
+            LOG.error("Node {} not found in native connection", nodeID);
+            return null;
+        }
         JMXConnector connector =  myJMXConnections.get(nodeID);
         if (isConnected(nodeID))
         {
@@ -127,12 +133,6 @@ public class DistributedJmxConnectionProviderImpl implements DistributedJmxConne
             return connector;
         }
         LOG.info("Connection expired or disconnected with node Id {}, attempting reconnection", nodeID);
-        Node node = myNativeConnectionProvider.getNodes().get(nodeID);
-        if (node == null)
-        {
-            LOG.error("Node {} not found in native connection provider", nodeID);
-            return null;
-        }
         try
         {
             LOG.info("Attempting to create JMX connection with node {}", node.getHostId());
